Data.Conduit.Filesystem

Description

NOTE It is recommended to start using Data.Conduit.Combinators instead of this module.

Documentation

sourceDirectory :: forall (m :: Type -> Type) i. MonadResource m => FilePath -> ConduitT i FilePath m ()

Stream the contents of the given directory, without traversing deeply.

This function will return all of the contents of the directory, whether they be files, directories, etc.

Note that the generated filepaths will be the complete path, not just the filename. In other words, if you have a directory foo containing files bar and baz, and you use sourceDirectory on foo, the results will be foo/bar and foo/baz.

Since: conduit-1.3.0

sourceDirectoryDeep

Arguments

:: forall (m :: Type -> Type) i. MonadResource m

=> Bool
Follow directory symlinks

-> FilePath
Root directory

-> ConduitT i FilePath m ()

Deeply stream the contents of the given directory.

This works the same as sourceDirectory, but will not return directories at all. This function also takes an extra parameter to indicate whether symlinks will be followed.

Since: conduit-1.3.0