Data.Conduit.Zlib

Description

Streaming compression and decompression using conduits.

Parts of this code were taken from zlib-enum and adapted for conduits.

Conduits

compress

Arguments

:: (PrimMonad m, MonadThrow m)

=> Int
Compression level

-> WindowBits
Zlib parameter (see the zlib-bindings package as well as the zlib C library)

-> ConduitT ByteString ByteString m ()

Compress (deflate) a stream of ByteStrings. The WindowBits also control the format (zlib vs. gzip).

decompress

Arguments

:: (PrimMonad m, MonadThrow m)

=> WindowBits
Zlib parameter (see the zlib-bindings package as well as the zlib C library)

-> ConduitT ByteString ByteString m ()

Decompress (inflate) a stream of ByteStrings. For example:

   sourceFile "test.z" $= decompress defaultWindowBits $$ sinkFile "test"

gzip :: (MonadThrow m, PrimMonad m) => ConduitT ByteString ByteString m ()

Gzip compression with default parameters.

ungzip :: (PrimMonad m, MonadThrow m) => ConduitT ByteString ByteString m ()

Gzip decompression with default parameters.

Flushing

compressFlush

Arguments

:: (PrimMonad m, MonadThrow m)

=> Int
Compression level

-> WindowBits
Zlib parameter (see the zlib-bindings package as well as the zlib C library)

-> ConduitT (Flush ByteString) (Flush ByteString) m ()

Same as compress, but allows you to explicitly flush the stream.

decompressFlush

Arguments

:: (PrimMonad m, MonadThrow m)

=> WindowBits
Zlib parameter (see the zlib-bindings package as well as the zlib C library)

-> ConduitT (Flush ByteString) (Flush ByteString) m ()

Same as decompress, but allows you to explicitly flush the stream.

Decompression combinators

multiple :: Monad m => ConduitT ByteString a m () -> ConduitT ByteString a m ()

The standard decompress and ungzip functions will only decompress a single compressed entity from the stream. This combinator will exhaust the stream completely of all individual compressed entities. This is useful for cases where you have a concatenated archive, e.g. cat file1.gz file2.gz > combined.gz.

Usage:

sourceFile "combined.gz" $$ multiple ungzip =$ consume

This combinator will not fail on an empty stream. If you want to ensure that at least one compressed entity in the stream exists, consider a usage such as:

sourceFile "combined.gz" $$ (ungzip >> multiple ungzip) =$ consume

Since: 1.1.10

Re-exported from zlib-bindings

data WindowBits

This specifies the size of the compression window. Larger values of this parameter result in better compression at the expense of higher memory usage.

The compression window size is the value of the the window bits raised to the power 2. The window bits must be in the range 9..15 which corresponds to compression window sizes of 512b to 32Kb. The default is 15 which is also the maximum size.

The total amount of memory used depends on the window bits and the MemoryLevel. See the MemoryLevel for the details.

Constructors

WindowBits Int

defaultWindowBits :: WindowBits

The default WindowBits is 15 which is also the maximum size.