Data.Conduit.Attoparsec

Description

Consume attoparsec parsers via conduit.

This code was taken from attoparsec-enumerator and adapted for conduits.

Sink

sinkParser :: (AttoparsecInput a, MonadThrow m) => Parser a b -> ConduitT a o m b

Convert an Attoparsec Parser into a Sink. The parser will be streamed bytes until it returns Done or Fail.

If parsing fails, a ParseError will be thrown with throwM.

Since 0.5.0

sinkParserEither :: (AttoparsecInput a, Monad m) => Parser a b -> ConduitT a o m (Either ParseError b)

Same as sinkParser, but we return an Either type instead of raising an exception.

Since 1.1.5

Conduit

conduitParser :: (AttoparsecInput a, MonadThrow m) => Parser a b -> ConduitT a (PositionRange, b) m ()

Consume a stream of parsed tokens, returning both the token and the position it appears at. This function will raise a ParseError on bad input.

Since 0.5.0

conduitParserEither :: (Monad m, AttoparsecInput a) => Parser a b -> ConduitT a (Either ParseError (PositionRange, b)) m ()

Same as conduitParser, but we return an Either type instead of raising an exception.

Types

data ParseError

The context and message from a Fail value.

Constructors

ParseError

Fields

errorContexts :: [String]

errorMessage :: String

errorPosition :: Position

DivergentParser

data Position

Constructors

Position

Fields

posLine :: !Int

posCol :: !Int

posOffset :: !Int

Since: 1.2.0

data PositionRange

Constructors

PositionRange

Fields

posRangeStart :: !Position

posRangeEnd :: !Position

Classes

class AttoparsecInput a

A class of types which may be consumed by an Attoparsec parser.

Minimal complete definition

parseA, feedA, empty, isNull, getLinesCols, stripFromEnd