Data.Conduit.Foldl

Description

Adapter module to work with the foldl package.

Since: 1.1.16

Documentation

sinkFold :: Monad m => (x -> a -> x) -> x -> (x -> b) -> Consumer a m b

Convert a left fold into a Consumer. This function is intended to be used with purely from the foldl package.

Since: 1.1.16

sinkFoldM :: Monad m => (x -> a -> m x) -> m x -> (x -> m b) -> Consumer a m b

Convert a monadic left fold into a Consumer. This function is intended to be used with impurely from the foldl package.

Since: 1.1.16