Data.Conduit.Network.Unix

Basic utilities

sourceSocket :: MonadIO m => Socket -> ConduitT i ByteString m ()

Stream data from the socket.

This function does not automatically close the socket.

Since 0.0.0

sinkSocket :: MonadIO m => Socket -> ConduitT ByteString o m ()

Stream data to the socket.

This function does not automatically close the socket.

Since 0.0.0

Simple server/client interface

data AppDataUnix

The data passed to a Unix domain sockets Application.

appSource :: (HasReadWrite ad, MonadIO m) => ad -> ConduitT i ByteString m ()

appSink :: (HasReadWrite ad, MonadIO m) => ad -> ConduitT ByteString o m ()

Server

data ServerSettingsUnix

Settings for a Unix domain sockets server.

serverSettings :: FilePath -> ServerSettingsUnix

runUnixServer :: ServerSettingsUnix -> (AppDataUnix -> IO ()) -> IO a

Run an Application with the given settings. This function will create a new listening socket, accept connections on it, and spawn a new thread for each connection.

Client

data ClientSettingsUnix

Settings for a Unix domain sockets client.

clientSettings :: FilePath -> ClientSettingsUnix

runUnixClient :: ClientSettingsUnix -> (AppDataUnix -> IO a) -> IO a

Run an Application by connecting to the specified server.

Getters

getPath :: HasPath a => a -> FilePath

getAfterBind :: HasAfterBind a => a -> Socket -> IO ()

Setters

setPath :: HasPath a => FilePath -> a -> a

setAfterBind :: HasAfterBind a => (Socket -> IO ()) -> a -> a