Data.Conduit.Process

Description

A full tutorial for this module is available on FP School of Haskell: https://www.fpcomplete.com/user/snoyberg/library-documentation/data-conduit-process.

Note that this is a very thin layer around the Data.Streaming.Process module. In particular, it:

Provides orphan instances for conduit

Provides some useful helper functions

Functions

sourceCmdWithConsumer :: MonadIO m => String -> Consumer ByteString m a -> m (ExitCode, a)

Like sourceProcessWithConsumer but providing the command to be run as a String.

Since 1.1.2

sourceProcessWithConsumer :: MonadIO m => CreateProcess -> Consumer ByteString m a -> m (ExitCode, a)

Given a CreateProcess, run the process, with its output being used as a Source to feed the provided Consumer. Once the process has completed, return a tuple of the ExitCode from the process and the output collected from the Consumer.

Since 1.1.2

Reexport

module Data.Streaming.Process